Add basic scaffolding
- Implement local settings
- Configure cache
- Setup our basic view and URLs

Not sure yet what approach is better:
- Unmanaged model with standard serializer / get_queryset views
- Using APIView and implementing a lot of logic that goes automatically
with models

The first approach is in effect a cache as well, but for this kind of
data saving every lon/lat combination in a database quickly becomes
unmanageable.

The second approach requires a bit more research, since well, we're
working against the flow of both DRF and Django here, so we may need to
override a lot of things.

Let's start with a serializer and view and see where we end up.