Implement top-down view:
So far we do not seem to need a model and can just drive everything from
APIView's get method.

To make the remote api backend pluggable, delegate some methods
properly.

Next up: an adapter for the remote data. Can we use a Serializer here or
just return a dict through a basic adapter interface. DRF's serializer
may get in the way here. Need to see what it accepts as input.