Implement an adapter for OSM
Turns out we can just use a serializer and it's not getting in the way.
Add an adapter we can soon get configured and use reduce to convert a
dotted path to dictionary keys so we can easily map our input fields to
constructor arguments for the Address class.

The adapter has one method to implement: get_address(). By default, add
an OSM adapter.

While in here, chop the trailing slash off of the endpoint URL and
properly align endpoint to start of string.